use Translatable concern instead of controller inheritance
