+++ app/controllers/spree/admin/products_controller.rb
@@ -1,7 +1,8 @@
 module Spree
   module Admin
-    class ProductsController < TranslatableResourceController
+    class ProductsController < ResourceController
       include Spree::Admin::ProductConcern
+      include Translatable
 
       helper 'spree/admin/products'
 
@@ -139,8 +140,6 @@ module Spree
 
         params[:q] ||= {}
         params[:q][:deleted_at_null] ||= '1'
-
-        # This doesnt work because name asc takes spree_product.name, it is needed to use order with i18n before.
         params[:q][:s] ||= 'name asc'
 
         @collection = product_scope
